spapr: use DEVICE_UNPLUG_ERROR to report unplug errors
Linux Kernel 5.12 is now unisolating CPU DRCs in the device_removal
error path, signalling that the hotunplug process were not successful.
This allow us to send a DEVICE_UNPLUG_ERROR in drc_unisolate_logical()
to signal this error to the management layer.

We also have another error path in spapr_memory_unplug_rollback()
for configured LMB DRCs. Kernels older than 5.13 will not unisolate
the LMBs in the hotunplug error path, but it will reconfigure them.
Let's also send the DEVICE_UNPLUG_ERROR event in that code path as
well.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>