target/ppc: PMU Book3s basic insns count for pseries TCG
The PMCC (PMC Control) bit in the MMCR0 register controls whether the
counters PMC5 and PMC6 are being part of the performance monitor
facility in a specific time. If PMCC allows it, PMC5 and PMC6 will
always be used to measure instructions completed and cycles,
respectively.

This patch adds the barebones of the Book3s PMU logic by enabling
instruction countingi, using the icount framework, using the performance
monitor counters 5 and 6. The overall logic goes as follows:

- we're storing the icount state of each register in the CPUPPCState,
allowing for the current instructions on read/write of the PMCs or in
a freeze counters state change;

- only PMC5 and PMC6 are being set. PMC6 (cycles) is default to 4*insns
(for cycles per instruction) for now;

- the only bit that is operational is MMCR0_FC (Freeze Counters). All
other bits or registers such as MMCR1 are not being read. PMCC bit
is also being ignored, meaning that PMC5 and PMC6 will always be enabled
for the PMU.

Since there will be more PMU exclusive code to be added next, let's also
put the PMU logic in its own helper to keep all in the same place. The
code is also repetitive and not really extensible to add more PMCs. This
will be handled in the next patch.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>