target/ppc/power8_pmu.c: enable counter negative for PM_CYC
All performance monitor counters can trigger a counter negative
condition if the proper MMCR0 bits are set. This patch does for
all PMCs that can count cycles by doing the following:

- pmc_counter_negative_enabled() will check whether a given PMC is
eligible to trigger the counter negative alert;

- get_counter_neg_timeout() will return the timeout for the counter
negative condition for a given PMC, or -1 if the PMC is not able to
trigger this alert;

- the existing counter_negative_cond_enabled() now must consider the
counter negative bit for PMCs 2-6, MMCR0_PMCjCE;

- set_PMU_excp_timer() will now search all existing PMCs for the
shortest counter negative timeout. The shortest timeout will be used to
set the PMC interrupt timer.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>