target/ppc: initialize PMUEvents on MMCR1 write
The value of MMCR1 determines the events that are going to be sampled by
the programmable counters (PMCs 1-4). PMCs 5 and 6 are always counting
instructions and cycles respectively and aren't affected by MMCR1.

This patch adds a helper to initialize PMCs 1-4 PMUEvents when writing the
MMCR1 register. The following events are considered valid at this
moment:

- For PMCs 1-4, event 0x2 is the implementation dependent value of
PMU_EVENT_INSTRUCTIONS and event 0x1E is the implementation dependent value of
PMU_EVENT_CYCLES. These events are supported by IBM Power chips since
Power8, at least, and the Linux Perf driver makes use of these events
until kernel v5.15;

- for PMC1, event 0xF0 is the architected PowerISA event for cycles. Event 0xFE
is the architected PowerISA event for instructions.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>