target/ppc: enhance error report in kvmppc_read_int_cpu_dt()
First and foremost, the function can't return '-1' when an error occurs
because the return type is set to uint64_t. Let's fix that.

After that, the function can't simply return 0 whether an error happened
and call it a day. We must provide a way of letting callers know if the
zero return is legitimate or due to an error.

Add an Error pointer to kvmppc_read_int_cpu_dt() that will be filled
with an appropriate error, if one occurs. Callers are then free to pass
an Error pointer and handle it.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>